implement balance transfer

+++ src/main/java/ru/smartech/app/resources/impl/PhoneManagementResourceImpl.java
@@ -4,27 +4,32 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
-import ru.smartech.app.dto.EmailDto;
 import ru.smartech.app.dto.PhoneDto;
-import ru.smartech.app.entity.Email;
 import ru.smartech.app.entity.Phone;
 import ru.smartech.app.entity.User;
 import ru.smartech.app.resources.PhoneManagementResource;
 import ru.smartech.app.service.PhoneService;
+import ru.smartech.app.service.SecurityService;
 
 @Slf4j
 @RestController
 public class PhoneManagementResourceImpl implements PhoneManagementResource {
 
     private final PhoneService phoneService;
+    private final SecurityService securityService;
 
     @Autowired
-    public PhoneManagementResourceImpl(PhoneService phoneService) {
+    public PhoneManagementResourceImpl(PhoneService phoneService,
+                                       SecurityService securityService) {
         this.phoneService = phoneService;
+        this.securityService = securityService;
     }
 
     @Override
     public ResponseEntity<PhoneDto> linkPhone(String phone, long userId) {
+        if (!securityService.isPrincipal(userId))
+            return ResponseEntity.badRequest().build();
+
         return ResponseEntity.ok(
                 PhoneDto.from(
                         phoneService.add(
@@ -38,6 +43,10 @@ public class PhoneManagementResourceImpl implements PhoneManagementResource {
 
     @Override
     public ResponseEntity<PhoneDto> updatePhone(long phoneId, String phone) {
+        var existed = phoneService.findById(phoneId);
+        if (existed.isEmpty() || !securityService.isPrincipal(existed.get().getUser().getId()))
+            return ResponseEntity.badRequest().build();
+
         return ResponseEntity.ok(
                 PhoneDto.from(
                         phoneService.update(
@@ -51,6 +60,9 @@ public class PhoneManagementResourceImpl implements PhoneManagementResource {
 
     @Override
     public ResponseEntity<PhoneDto> unlinkPhone(long phoneId, long userId) {
+        if (!securityService.isPrincipal(userId))
+            return ResponseEntity.badRequest().build();
+
         phoneService.delete(
                 new Phone()
                         .setId(phoneId)
